pyproject.toml
[project]
name = "voice-backend"
version = "0.1.0"
description = "API de síntesis y clonación de voz usando ResembleAI/chatterbox-turbo"
authors = [
    {name = "Agus", email = "agus@example.com"}
]
dependencies = [
    "fastapi",
    "uvicorn",
    "python-multipart",
    "python-dotenv",
    "requests",
    "huggingface_hub",
    "chatterbox-tts>=0.1.6",
    "torch",
    "torchaudio",
    "numpy<2.0",
    "accelerate",
    "scipy",
    "sqlalchemy",
    "psycopg2-binary",
    "bcrypt",
    "python-jose[cryptography]",
    "alembic",
    "mlx-lm>=0.29.1",
    "mlx>=0.30.3",
    "openai-whisper>=20250625",
    "setuptools-rust>=1.12.0",
    "edge-tts>=7.2.7",
    "mcp>=1.25.0",
]
requires-python = ">=3.11, <3.12"
readme = "README.md"
license = {text = "MIT"}

[tool.pdm]
distribution = false

[tool.pdm.scripts]
dev = { cmd = "uvicorn main:app --reload --host 0.0.0.0 --port 8000", env = { ENVIRONMENT = "development" } }
start = { cmd = "python main.py", env = { ENVIRONMENT = "production" } }
makemigrations = "alembic revision --autogenerate -m"
migrate = "alembic upgrade head"
